refactor(perks): split apply handlers into per-perk modules
