+++ src/crimson/perks/death_clock.py
@@ -1,0 +1,28 @@
+from __future__ import annotations
+
+from ..sim.state_types import PlayerState
+from .apply_context import PerkApplyCtx
+from .helpers import perk_count_get
+from .ids import PerkId
+
+
+def _increment_perk_count(player: PlayerState, perk_id: PerkId, *, amount: int = 1) -> None:
+    idx = int(perk_id)
+    if 0 <= idx < len(player.perk_counts):
+        player.perk_counts[idx] += int(amount)
+
+
+def apply_death_clock(ctx: PerkApplyCtx) -> None:
+    _increment_perk_count(
+        ctx.owner,
+        PerkId.REGENERATION,
+        amount=-perk_count_get(ctx.owner, PerkId.REGENERATION),
+    )
+    _increment_perk_count(
+        ctx.owner,
+        PerkId.GREATER_REGENERATION,
+        amount=-perk_count_get(ctx.owner, PerkId.GREATER_REGENERATION),
+    )
+    for player in ctx.players:
+        if player.health > 0.0:
+            player.health = 100.0
