+++ src/crimson/perks/infernal_contract.py
@@ -1,0 +1,13 @@
+from __future__ import annotations
+
+from .apply_context import PerkApplyCtx
+
+
+def apply_infernal_contract(ctx: PerkApplyCtx) -> None:
+    ctx.owner.level += 3
+    if ctx.perk_state is not None:
+        ctx.perk_state.pending_count += 3
+        ctx.perk_state.choices_dirty = True
+    for player in ctx.players:
+        if player.health > 0.0:
+            player.health = 0.1
